+++ RandLib/CategoricalDist/CategoricalDist.cs
@@ -3,18 +3,18 @@ namespace RandLib.CategoricalDist
     public class CategoricalDist
     {
         //gets a random number between 0 and 1 then subtracts it from pHeads. True if nonnegative
-        public static bool flip(float pHeads)
+        public static bool flip(double pHeads)
         {
             Random rng = new Random();
             double num = rng.NextDouble();
-            return ( (double)pHeads - num) >= 0;
+            return (pHeads - num) >= 0;
         }
         //seeded version of above, returns modified seed
-        public static (bool, int) flip(float pHeads, int seed)
+        public static (bool, int) flip(double pHeads, int seed)
         {
             Random rng = new Random(seed);
             double num = rng.NextDouble();
-            return ( (double)pHeads >= num, rng.Next());
+            return ( pHeads >= num, rng.Next());
         }
     }
 }
